Rework STATUS_VIA_WINDOWPORT
This simplifies the core-to-window-port API for updating the status area
by sending all status info every time, instead of piecemeal when
individual status fields change.  This makes the status update logic
much easier to reason about, as well as shedding a substantial amount of
book-keeping and generic parameter handling that the old system had to
do.

This also removes the status delta logic, which will return in a
different form when the statuscolors feature is implemented.

Currently only tested for the TTY window port.