Rewrite hitpointbar patch for NetHack 3.6.0
This shows a graphical bar over the name and title in the status area
representing remaining hit points.  If statuscolors are compiled in,
enabled and set for HP, those colors will be shared for the bar,
otherwise the bar will simply display in inverse.

To try hitpointbar, add STATUS_VIA_WINDOWPORT to your CFLAGS, compile
the game, then enable the "hitpointbar" boolean option (it's off by
default).

Currently only implemented for TTY, but the hitpointbar setting is
designed as a window capability, so other window ports may choose
implement it in the future.