Add statuscolors and hitpointbar to Guidebook
Guidebook.mn and Guidebook.tex specifically, for reasons stated earlier.

This adds the statuscolors and hitpointbar boolean options, as well as a
"Configuring Status Colors" section to Guidebook.mn and Guidebook.tex.

I cargo-culted the hell out of those document markup syntaxes, so let's
hope I got it right, or close enough to that fixing things won't be too
hard.